move FileUserRepository out of preferences
